fix: made q6k work as expected
